Tweak changeset 5b8c91e4183664c90f4d95699ab971e674ffd67d which created symlinks to a required tool dep endncy env.sh file for dependent tool dependencies being installed.  This approach was better than the previous approach which parsed the requried repository's tool_dependencies.xml file to create the env.sh file.  This changeset uses a slightly different approach: it creates an env.sh file for the dependent tool dependency whose first line "sources" the required tool dende pendcy's env.sh file.  This allows for the dependent env.sh file to be managed by its tool dependency (e.g., its tool dependency may require additional env.sh paths beyond those contained in the requyired tool dependency's env.sh file).  This changeset also handles sysncing the database with the file system when tool dependencies are being installed when handling a complex repository dependency definition.  This process used to be handled only when not dealing with a complex repository dependency definition.  This fix should correct problems encounted by the Tool Shed's install and test framework which were introduced in 5b8c91e4183664c90f4d95699ab971e674ffd67d.

+++ lib/tool_shed/util/shed_util_common.py
@@ -1521,11 +1521,21 @@ def pretty_print( dict=None ):
 
 def remove_dir( dir ):
     """Attempt to remove a directory from disk."""
-    if os.path.exists( dir ):
-        try:
-            shutil.rmtree( dir )
-        except:
-            pass
+    if dir:
+        if os.path.exists( dir ):
+            try:
+                shutil.rmtree( dir )
+            except:
+                pass
+
+def remove_file( file_name ):
+    """Attempt to remove a file from disk."""
+    if file_name:
+        if os.path.exists( file_name ):
+            try:
+                os.remove( dir )
+            except:
+                pass
 
 def repository_was_previously_installed( trans, tool_shed_url, repository_name, repo_info_tuple ):
     """
